Extending TiffImages.jl

If you want to extend TiffImages.jl to add support for more features or change how TIFF data is loaded, you have come to right place.

Types

TiffImages.TiffFile — Type

mutable struct TiffFile{O<:Unsigned, S<:FileIO.Stream}

-> TiffFile

Wrap io with helper parameters to keep track of file attributes.

filepath

The relative path to this file

io

The file stream

first_offset

Location of the first IFD in the file stream

need_bswap

Whether this file has a different endianness than the host computer

TiffImages.IFD — Type

struct IFD{O<:Unsigned}

An image file directory is a sorted collection of the tags representing this plane in the TIFF file. They behave like dictionaries, so given an IFD called ifd, we can add new tags as follows:

julia> ifd[TiffImages.IMAGEDESCRIPTION] = "Some details";

julia> ifd[TiffImages.IMAGEWIDTH] = 512;

julia> ifd
IFD, with tags:
	Tag(IMAGEWIDTH, 512)
	Tag(IMAGEDESCRIPTION, "Some details")

TiffImages.Tag — Type

struct Tag{T}

In-memory representation of Tiff Tags, which are essentially key value pairs. The data field can either be a String, a Number, an Array of bitstypes, or a RemoteData type.

tag

data

TiffImages.RemoteData — Type

RemoteData

A placeholder type to describe the location and properties of remote data that is too large to fit directly in a tag's spot in the IFD. Calling load! on an IFD object replaces all RemoteDatas with the respective data.

position

Position of this data in the stream

datatype

The datatype stored at this location

count

The length of the data